♻️ refactor: 파스칼 케이스로 ts 이름 변경
