+++ src/constants/globalConstants.ts
